+++ packages/service-configuration/service_configuration_server.js
@@ -3,7 +3,7 @@
 // otherwise lead to an inconsistent database state (when there are multiple
 // configurations for a single service, which configuration is correct?)
 try {
-    ServiceConfiguration.configurations._ensureIndex(
+    ServiceConfiguration.configurations.createIndex(
         { "service": 1 },
         { unique: true }
     );
@@ -14,7 +14,7 @@ try {
         "each service should have exactly one configuration, Meteor " +
         "automatically creates a MongoDB index with a unique constraint on the " +
         " meteor_accounts_loginServiceConfiguration collection. The " +
-        "_ensureIndex command which creates that index is failing.\n\n" +
+        "createIndex command which creates that index is failing.\n\n" +
         "Meteor versions before 1.0.4 did not create this index. If you recently " +
         "upgraded and are seeing this error message for the first time, please " +
         "check your meteor_accounts_loginServiceConfiguration collection for " +
@@ -22,7 +22,7 @@ try {
         "configuration entries until there is no more than one configuration " +
         "entry per service.\n\n" +
         "If the meteor_accounts_loginServiceConfiguration collection looks " +
-        "fine, the _ensureIndex command is failing for some other reason.\n\n" +
+        "fine, the createIndex command is failing for some other reason.\n\n" +
         "For more information on this history of this issue, please see " +
         "https://github.com/meteor/meteor/pull/3514.\n"
     );